% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{add_params}
\alias{add_params}
\title{Params}
\arguments{
\item{serie}{column of values to plot.}

\item{name}{name of serie.}

\item{stack}{set to stack.}

\item{axis}{axis to plot on, see examples.}
}
\description{
Params
}
\keyword{internal}
