% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_xaxis_label}
\alias{b_xaxis_label}
\alias{b_yaxis_label}
\title{Add axis labels}
\usage{
b_xaxis_label(p, label, position = "inner-top")

b_yaxis_label(p, label, position = "inner-top", axis = "y")
}
\arguments{
\item{p}{billboard object as returned by \code{\link{b_board}}.}

\item{label}{axis label.}

\item{position}{position of label.}

\item{axis}{axis to add label to.}
}
\description{
Add axis labels.
}
\details{
Valid \code{position}:
\itemize{
  \item{\code{inner-top}}
  \item{\code{inner-middle}}
  \item{\code{inner-bottom}}
  \item{\code{outer-top}}
  \item{\code{outer-middle}}
  \item{\code{outer-bottom}}
}
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_spline(cyl) \%>\%
  b_xaxis_label("Weight")

mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") \%>\%
  b_xaxis_label("Model") \%>\%
  b_yaxis_label("Miles per galon") \%>\%
  b_yaxis_label("Rear axle ratio", axis = "y2")

}
