% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_area}
\alias{b_area}
\title{Add area}
\usage{
b_area(p, serie, name = NULL, stack = FALSE, axis = "y",
  zerobased = TRUE, above = FALSE)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{serie}{column of values to plot.}

\item{name}{name of serie.}

\item{stack}{set to stack.}

\item{axis}{axis to plot on, see examples.}

\item{zerobased}{set if min or max value will be 0 on bar chart.}

\item{above}{set to fill area above.}
}
\description{
Add an area.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_area(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_area(mpg, stack = TRUE) \%>\%
  b_area(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_area(mpg) \%>\%
  b_area(drat, axis = "y2")

}
