% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_bar}
\alias{b_bar}
\title{add bar}
\usage{
b_bar(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\description{
add bar
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_bar(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_bar(mpg, stack = TRUE) \%>\%
  b_bar(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_bar(mpg) \%>\%
  b_bar(drat, axis = "y2")

}
