% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard.R
\name{b_board}
\alias{b_board}
\title{Create billboard object}
\usage{
b_board(data, x = NULL, width = "100\%", height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{data.frame containing data to plot.}

\item{x}{variable column.}

\item{width, height}{dimensions of chart.}

\item{elementId}{id of div containing chart.}
}
\description{
Create billboard chart.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_line(wt)

}
