% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_donut}
\alias{b_donut}
\title{Add donut}
\usage{
b_donut(p, serie, label = list(), expand = TRUE, width = NULL,
  title = "", pad.angle = 0)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{serie}{name of serie to plot.}

\item{label}{label parameters as list.}

\item{expand}{set to expand on hover.}

\item{width}{width of donut.}

\item{title}{donut title}

\item{pad.angle}{set padding between data.}
}
\description{
Add a donut (chart).
}
\examples{
library(dplyr)

mtcars \%>\%
  mutate(model = rownames(.)) \%>\%
  slice(1:5) \%>\%
  b_board(model) \%>\%
  b_donut(mpg)

}
