% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_inter}
\alias{b_inter}
\title{Chart interactions}
\usage{
b_inter(p, enabled = TRUE, brighten = TRUE, ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{enabled}{set to enable.}

\item{brighten}{set to brighten.}

\item{...}{any other parameter.}
}
\description{
Chart interactions.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_area(mpg) \%>\%
  b_inter(brighten = FALSE)

}
