% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_line}
\alias{b_line}
\title{add line}
\usage{
b_line(p, serie, name = NULL, stack = FALSE)
}
\description{
add line
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_line(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_line(mpg, stack = TRUE) \%>\%
  b_line(drat, stack = TRUE)

}
