% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{b_lines}
\alias{b_lines}
\title{Customise line}
\usage{
b_lines(p, connect = FALSE, step = NULL, ...)
}
\description{
Customise line
}
\examples{
df <- data.frame(x = 1:20, y = c(3, NA, runif(18, 1, 5)))

df \%>\%
  b_board(x) \%>\%
  b_step(y) \%>\%
  b_lines(connect = TRUE, step = "step-after")

}
