% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{b_point}
\alias{b_point}
\title{Customise point}
\usage{
b_point(p, show = TRUE, r = 2.5, expand.enable = TRUE, expand.r = 4.5,
  select.r = 4, ...)
}
\description{
Customise point
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_line(disp) \%>\%
  b_point(r = 4)

}
