% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_region}
\alias{b_region}
\title{Add regions}
\usage{
b_region(p, axis, start, end, class = NULL, ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{axis}{axis to draw region on.}

\item{start, end}{start and end of region.}

\item{class}{class of region (\code{CSS}).}

\item{...}{any other parameter.}
}
\description{
Add regions to chart.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_scatter(mpg) \%>\%
  b_region(axis = "x", start = 8, end = 20) \%>\%
  b_region(axis = "x", start = 1, end = 4) \%>\%
  b_region(axis = "y", start = 30, end = 34)

}
