% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_resize}
\alias{b_resize}
\title{Resize}
\usage{
b_resize(p, auto = TRUE)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{auto}{set to automatically resize chart.}
}
\description{
Chart resize.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_scatter(mpg) \%>\%
  b_resize(FALSE)

}
