% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_rotate}
\alias{b_rotate}
\title{Rotate axis}
\usage{
b_rotate(p)
}
\arguments{
\item{p}{an object of class \code{billboard}.}
}
\description{
Rotate axis
}
\examples{
mtcars[1:5,] \%>\%
  b_board(wt) \%>\%
  b_bar(drat) \%>\%
  b_rotate()

}
