% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_scatter}
\alias{b_scatter}
\title{add scatter}
\usage{
b_scatter(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\description{
add scatter
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_scatter(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_scatter(mpg, stack = TRUE) \%>\%
  b_scatter(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_scatter(mpg) \%>\%
  b_scatter(drat, axis = "y2")

}
