% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_setx}
\alias{b_setx}
\title{Define X axis}
\usage{
b_setx(p)
}
\arguments{
\item{p}{an object of class \code{billboard}.}
}
\description{
Define X axis, pipe it after all series.
}
\details{
pipe if after all series.
}
\examples{
library(dplyr)

# this makes no sense, it's just an example.
mtcars \%>\%
  #' mtcars \%>\%
  group_by(wt) \%>\%
  summarise(
    mpg = sum(mpg),
    drat = mean(drat)
  ) \%>\%  
  b_board(wt) \%>\%
  b_line(mpg) \%>\%
  b_line(drat) \%>\%
  b_setx()
  
}
