% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_size}
\alias{b_size}
\title{Size}
\usage{
b_size(p, width = NULL, height = NULL)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{width, height}{dimensions.}
}
\description{
Chart size.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_scatter(mpg) \%>\%
  b_size(900, 500)

}
