% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_spline}
\alias{b_spline}
\title{add spline}
\usage{
b_spline(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\description{
add spline
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg, stack = TRUE) \%>\%
  b_spline(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2")

}
