% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_step}
\alias{b_step}
\title{add step}
\usage{
b_step(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\description{
add step
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_step(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_step(mpg, stack = TRUE) \%>\%
  b_step(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_step(mpg) \%>\%
  b_step(drat, axis = "y2")

}
