% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_svg}
\alias{b_svg}
\title{Add svg class}
\usage{
b_svg(p, class)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{class}{\code{CSS class}.}
}
\description{
Add class to svg.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_scatter(mpg) \%>\%
  b_svg("svg-class")

}
