% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_title}
\alias{b_title}
\title{Add title}
\usage{
b_title(p, title = NULL, position = "top-center", ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{title}{chart title.}

\item{position}{title position.}

\item{...}{any other parameter.}
}
\description{
Add a title to the chart.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_scatter(mpg) \%>\%
  b_title("Chart title", position = "top-right")

}
