% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_tooltip}
\alias{b_tooltip}
\title{Add tooltip}
\usage{
b_tooltip(p, show = TRUE, grouped = FALSE, order = "asc", ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{show}{set to show tooltip.}

\item{grouped}{set if tooltip is grouped or not for the data points.}

\item{order}{set tooltip data display order.}

\item{...}{any other parameter.}
}
\description{
Add a tooltip.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") \%>\%
  b_tooltip(grouped = TRUE) \%>\%
  b_xgrid(show = FALSE) \%>\%
  b_grid_line(14, "x = 14")

}
