% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_trans}
\alias{b_trans}
\title{Cutomise transition}
\usage{
b_trans(p, duration = 350)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{duration}{duration of animations.}
}
\description{
Set duration of transition for chart animation.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_scatter(mpg) \%>\%
  b_trans(1000)

}
