% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_xaxis}
\alias{b_xaxis}
\title{Customise axis}
\usage{
b_xaxis(p, show = TRUE, height = NULL, min = NULL, max = NULL,
  padding = NULL, ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{show}{set to show axis.}

\item{height}{height of axis.}

\item{min, max}{minimum and maximum.}

\item{padding}{axis padding.}

\item{...}{any other parameter.}
}
\description{
Customise charts' axis.
}
\examples{
mtcars \%>\%
  b_board(drat) \%>\%
  b_bar(cyl) \%>\%
  b_xlabel("Rear Axle ratio") \%>\%
  b_ylabel("Number of cylinders") \%>\%
  b_xaxis(padding = list(left = 1, right = 1))

}
