% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_xtick}
\alias{b_xtick}
\title{Customise X axis ticks}
\usage{
b_xtick(p, centered = TRUE, count = NULL, culling = NULL, fit = TRUE,
  multiline = TRUE, outer = TRUE, rotate = 0, width = NULL, ...)
}
\description{
Customise X axis ticks
}
\examples{
library(dplyr)

mtcars \%>\%
  mutate(model = rownames(.)) \%>\%
  b_board(model) \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") \%>\%
  b_xlabel("Model") \%>\%
  b_ylabel("Miles per galon") \%>\%
  b_ylabel("Rear axle ratio", axis = "y2") \%>\%
  b_xaxis(height = 130) \%>\%
  b_xtick(rotate = 90, multiline = FALSE)

}
