% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_ytick}
\alias{b_ytick}
\title{Customise Y axis ticks}
\usage{
b_ytick(p, axis = "y", count = NULL, outer = TRUE, ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{axis}{target axis.}

\item{count}{set to count.}

\item{outer}{set to position outside chart.}

\item{...}{any other parameter.}
}
\description{
Customise Y axis ticks.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") \%>\%
  b_ytick(count = 10) \%>\%
  b_ytick("y2", outer = FALSE)

}
