% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{b_zoom_p}
\alias{b_zoom_p}
\title{Zoom}
\usage{
b_zoom_p(chartProxy, domain)
}
\arguments{
\item{domain}{domain to zoom to.}

\item{an}{object of class \code{billboardProxy} as returned by \code{\link{billoardProxy}}.}
}
\description{
Zoom
}
\examples{
\dontrun{
library(shiny)
ui <- fluidPage(
  sidebarLayout(
    sliderInput("zoom",
      "Zoom on a region",
      min = 0,
      max = 100,
      value = 100
    ),
    mainPanel(
      billboardOutput("billboard")
    )
  )
)

server <- function(input, output){
  data <- data.frame(x = 1:100, y = runif(100, 1, 100))

  output$billboard <- renderBillboard({
    data \%>\% 
      b_board(x) \%>\% 
      b_line(y)
  })
  
  observeEvent(input$zoom, {
    billboardProxy("billboard") \%>\% 
    b_zoom_p(c(0, input$zoom))
  })
}

shinyApp(ui, server)
}

}
