% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{b_colors}
\alias{b_colors}
\alias{b_brewer}
\alias{b_viridis}
\alias{b_plasma}
\alias{b_inferno}
\alias{b_magma}
\title{Customise colors}
\usage{
b_colors(p, colors, ...)

b_brewer(p, pal, n = 8, ...)

b_viridis(p, ...)

b_plasma(p, ...)

b_inferno(p, ...)

b_magma(p, ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{colors}{colors to use.}

\item{...}{any other parameter.}

\item{pal}{\code{RColorBrewer} color palette.}

\item{n}{number of colors.}
}
\description{
Customise colors.
}
\examples{
mtcars \%>\%
  b_board(mpg) \%>\%
  b_area(qsec) \%>\%
  b_spline(wt) \%>\%
  b_bar(disp, axis = "y2") \%>\%
  b_step(cyl) -> p

p \%>\%
  b_colors(c("red", "blue", "yellow", "orange"))

p \%>\%
  b_brewer("Dark2")

p \%>\%
  b_viridis()

p \%>\%
  b_magma()

p \%>\%
  b_plasma()

}
