% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{b_grid}
\alias{b_grid}
\alias{b_xgrid}
\alias{b_ygrid}
\alias{b_grid_line}
\title{Customise grid}
\usage{
b_grid(p, focus = TRUE, lines = TRUE)

b_xgrid(p, show = TRUE, ...)

b_ygrid(p, show = TRUE, ...)

b_grid_line(p, value, label, axis = "x", position = NULL, class = NULL)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{focus}{show grids when focus.}

\item{lines}{show lines.}

\item{show}{set to show.}

\item{...}{any other parameter.}

\item{value}{axis value.}

\item{label}{label of line.}

\item{axis}{axis to draw on.}

\item{position}{line position.}

\item{class}{\code{CSS} class.}
}
\description{
Add and customise grid.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_bar(disp) \%>\%
  b_bar(hp, axis = "y2") -> p

# add grid
p \%>\%
  b_xgrid() \%>\%
  b_ygrid() -> p

p

# add lines
p \%>\%
  b_grid_line(8, "x = 8") \%>\%
  b_grid_line(25, "x = 25") \%>\%
  b_grid_line(max(mtcars$disp), "Max disp", axis = "y")

}
