% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lyrics}
\alias{lyrics}
\title{Lyrics from songs on Billboards Hot 100 from 1960 to 2016}
\format{A data frame with 5701 rows and 4 variables:

\describe{
  \item{title}{the title of the song}
  \item{artist}{the artist of the song}
  \item{year}{year}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{http://www.genius.com/}
}
\usage{
lyrics
}
\description{
A data set containing lyrics for songs on the Billboard Hot 100 over the past
57 years. The lyrics were identified and collected by webscraping so there
might be some errors and mistakes - have that in mind.
}
\examples{
if (require("dplyr")) {

  data(lyrics)

  lyrics \%>\% glimpse()

}
}
\keyword{datasets}
