% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_hot_100s}
\alias{wiki_hot_100s}
\title{Overview of Billboards Hot 100 from 1960 to 2016}
\format{A data frame with 5701 rows and 4 variables:

\describe{
  \item{no}{the rank that the song had that year}
  \item{title}{the title of the song}
  \item{artist}{the artist of the song}
  \item{year}{year}
}}
\source{
\url{http://www.wikipedia.com/}
}
\usage{
wiki_hot_100s
}
\description{
A data set containing 57 years of Billboards Hot 100 songs. The data is scraped
from Wikipedia from the urls 'https://en.wikipedia.org/wiki/Billboard_Year-End_Hot_100_singles_of_'
and then the year added. Example: https://en.wikipedia.org/wiki/Billboard_Year-End_Hot_100_singles_of_1960.
One year has more than a 100 songs due to a tie.
}
\examples{
if (require("dplyr")) {

  data(wiki_hot_100s)

  wiki_hot_100s \%>\% glimpse()

}
}
\keyword{datasets}
