% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_area}
\alias{b_area}
\title{add area}
\usage{
b_area(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\description{
add area
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_area(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_area(mpg, stack = TRUE) \%>\%
  b_area(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_area(mpg) \%>\%
  b_area(drat, axis = "y2")

}
