% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_donut}
\alias{b_donut}
\title{add donut}
\usage{
b_donut(p, serie)
}
\description{
add donut
}
\examples{
library(dplyr)

mtcars \%>\%
  mutate(model = rownames(.)) \%>\%
  slice(1:5) \%>\%
  b_board(model) \%>\%
  b_donut(mpg)

}
