% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{b_legend}
\alias{b_legend}
\title{Customise legend}
\usage{
b_legend(p, show = TRUE, position = "bottom", equally = FALSE,
  padding = 0, width = 10, height = 10, ...)
}
\description{
Customise legend
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") -> p

p \%>\%
  b_legend(width = 25, height = 20)

p \%>\%
  b_legend(FALSE)

}
