% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_line}
\alias{b_line}
\title{Add line}
\usage{
b_line(p, serie, name = NULL, stack = FALSE, axis = "y",
  connect = FALSE)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{serie}{column of values to plot.}

\item{name}{name of serie.}

\item{stack}{set to stack.}

\item{axis}{axis to plot on, see examples.}

\item{connect}{set if null data point will be connected or not.}
}
\description{
Add a line.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_line(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_line(mpg, stack = TRUE) \%>\%
  b_line(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_line(mpg) \%>\%
  b_line(drat, axis = "y2")

}
