% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_pie}
\alias{b_pie}
\title{add pie}
\usage{
b_pie(p, serie)
}
\description{
add pie
}
\examples{
library(dplyr)

mtcars \%>\%
  mutate(model = rownames(.)) \%>\%
  slice(1:5) \%>\%
  b_board(model) \%>\%
  b_pie(mpg)

}
