% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_spline}
\alias{b_spline}
\title{Add spline}
\usage{
b_spline(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{serie}{column of values to plot.}

\item{name}{name of serie.}

\item{stack}{set to stack.}

\item{axis}{axis to plot on, see examples.}
}
\description{
Add a spline.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg, stack = TRUE) \%>\%
  b_spline(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2")

}
