% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{b_step}
\alias{b_step}
\title{Add step}
\usage{
b_step(p, serie, name = NULL, stack = FALSE, axis = "y")
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{serie}{column of values to plot.}

\item{name}{name of serie.}

\item{stack}{set to stack.}

\item{axis}{axis to plot on, see examples.}
}
\description{
Add a step chart.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_step(mpg)

mtcars \%>\%
  b_board() \%>\%
  b_step(mpg, stack = TRUE) \%>\%
  b_step(drat, stack = TRUE)

mtcars \%>\%
  b_board() \%>\%
  b_step(mpg) \%>\%
  b_step(drat, axis = "y2")

}
