% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{b_tooltip}
\alias{b_tooltip}
\title{Add tooltip}
\usage{
b_tooltip(p, show = TRUE, grouped = FALSE, order = "asc", ...)
}
\description{
Add tooltip
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") \%>\%
  b_tooltip(grouped = TRUE)

}
