% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_xaxis}
\alias{b_xaxis}
\title{Customise axis}
\usage{
b_xaxis(p, show = TRUE, height = NULL, min = NULL, max = NULL,
  padding = NULL, ...)
}
\description{
Customise charts' axis.
}
\examples{
mtcars \%>\%
  b_board(wt) \%>\%
  b_spline(cyl) \%>\%
  b_xlabel("Weight") \%>\%
  b_ylabel("Number of cylinders") \%>\%
  b_xaxis(padding = list(left = 1, right = 1))

}
