% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{b_yaxis}
\alias{b_yaxis}
\title{Customise y axis}
\usage{
b_yaxis(p, axis = "y", show = TRUE, center = NULL, inner = FALSE,
  inverted = FALSE, min = NULL, max = NULL, ...)
}
\description{
Cutomise y and y2 axis.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_spline(mpg) \%>\%
  b_spline(drat, axis = "y2") \%>\%
  b_xlabel("Model") \%>\%
  b_ylabel("Miles per galon") \%>\%
  b_ylabel("Rear axle ratio", axis = "y2") \%>\%
  b_yaxis("y2", inner = TRUE) \%>\%
  b_xaxis(padding = list(left = 1, right = 2)) # add padding to avoid overlap

}
