% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{b_color}
\alias{b_color}
\alias{b_color_brewer}
\alias{b_color_viridis}
\alias{b_color_plasma}
\alias{b_color_inferno}
\alias{b_color_magma}
\title{Customise colors}
\usage{
b_color(p, colors, ...)

b_color_brewer(p, pal, n = 8, ...)

b_color_viridis(p, ...)

b_color_plasma(p, ...)

b_color_inferno(p, ...)

b_color_magma(p, ...)
}
\arguments{
\item{p}{an object of class \code{billboard}.}

\item{colors}{colors to use.}

\item{...}{any other parameter.}

\item{pal}{\code{RColorBrewer} color palette.}

\item{n}{number of colors.}
}
\description{
Customise colors.
}
\examples{
mtcars \%>\%
  b_board() \%>\%
  b_area(qsec) \%>\%
  b_spline(wt) \%>\%
  b_bar(disp, axis = "y2") \%>\%
  b_step(cyl) -> p

p \%>\%
  b_color(c("red", "blue", "yellow", "orange"))

p \%>\%
  b_color_brewer("Dark2")

p \%>\%
  b_color_viridis()

p \%>\%
  b_color_magma()

p \%>\%
  b_color_plasma()

}
